+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/json_web_key.hpp
@@ -37,7 +37,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      * @brief They type of the key.
      *
      */
-    JsonWebKeyType KeyType;
+    KeyVaultKeyType KeyType;
 
     /**
      * @brief Construct a new Json Web Key object.